Updating to Lara 11

+++ database/migrations/2022_02_14_023228_create_words_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->increments('id');
             $table->string('word');
             $table->float('score');
-            $table->float('letter_frequency_score');
+            $table->decimal('letter_frequency_score', 8, 5);
         });
     }
